Fix simple setting hole threhsold step

+++ components/calibration/simple/HoleSettings.tsx
@@ -21,7 +21,7 @@ type Props = {
 const HoleSettings = ({ dictionary, settings, onSettingsChange }: Props) => {
   const onChange = (field: string) => {
     return (event: React.ChangeEvent<HTMLInputElement>) => {
-      const value = Number.parseInt(event.target.value);
+      const value = Number.parseFloat(event.target.value);
       const stepSettings = {
         ...settings[StepName.EXTRACT_OBJECT],
         holeSettings: {
